requirements.txt
# Core dependencies
anthropic>=0.40.0
requests>=2.31.0
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.2
beautifulsoup4>=4.12.0
jinja2>=3.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Type hints
typing-extensions>=4.12.0
